scripts/rules_builder/requirements.txt
ipython
jinja2==3.0.3
PyYAML==6.0
pytest
